% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cointro_network_graph.R
\name{cointro_network_graph}
\alias{cointro_network_graph}
\title{Create Network Graph of Cointroductions of Honolulu City Council Measures}
\usage{
cointro_network_graph(measures_df, type = "all")
}
\arguments{
\item{measures_df}{The measures tibble data frame created by `get_council_measures`}

\item{type}{The type of measure to selection. Options are "all", "bill", or "resolution".}
}
\value{
igraph network plot
}
\description{
Create Network Graph of Cointroductions of Honolulu City Council Measures
}
\details{
This function creates a network plot using igraph that visualizes how often
Honolulu City Council members introduce measures together. Each line in the
graph indicates one measure that the members introduced together. Honolulu
City Council measures can only be introduced by a max of two members, though
many measures only have one introducer. Measures that were only introduced by
one member are not shown in the graph.
}
\examples{
measures_df <- get_council_measures(start_measure = 2796, end_measure = 2797)
cointro_network_graph(measures_df, type = "all")
}
