% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_council_measures.R
\name{get_council_measures}
\alias{get_council_measures}
\title{Download Information About Honolulu City Council Measures}
\usage{
get_council_measures(start_measure = 2796, end_measure = 3162)
}
\arguments{
\item{start_measure}{The numeric value of the starting measure.}

\item{end_measure}{The numeric value of the ending measure}
}
\value{
A tibble data frame with information about each measure.
}
\description{
`get_council_measures` returns a data frame with information about all
measures that occur within the specified range of measures.
}
\details{
This function downloads information from the Honolulu City Council's website
about all measures that occur within the range from the start_measure to the
end_measure. Those values must be numeric. If only a single measure is wanted,
then set the start_measure and end_measure to the same value. The numeric
values of the measures differ from the measures' names. For example, RES25-001
has a numeric value of 3163, since that is the number given in its url:
https://hnldoc.ehawaii.gov/hnldoc/measure/3163. The values appear to occur in
sequential order, though there are some occasional exceptions.
}
\examples{
measures_df <- get_council_measures(start_measure = 2796, end_measure = 2797)

}
