% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_votes.R
\name{get_votes}
\alias{get_votes}
\title{Extract Voting Results for Honolulu City Council Measures}
\usage{
get_votes(measures)
}
\arguments{
\item{measures}{A tibble data frame of measures downloaded by `get_council_measures`}
}
\value{
A tibble data frame in long format with one row per council member per voting event.
}
\description{
Extract Voting Results for Honolulu City Council Measures
}
\details{
This function creates a long tibble data frame with one row per voting event per
Honolulu City Council member who participated in the voting event. The output of
this function excludes non-voting events such as the introduction of a measure
or publishing public hearing notices.
}
\examples{
measures_df <- get_council_measures(start_measure = 2796, end_measure = 2797)
votes_df <- get_votes(measures_df)
}
